+++ src/Model/BaseModel.php
@@ -1,0 +1,9 @@
+<?php
+
+namespace Exist404\DatatableCruds\Model;
+
+use Illuminate\Database\Eloquent\Model;
+
+class BaseModel extends Model
+{
+}